Removed PHP Symlinks.
Previously, there was a symlink in the repository that pointed to the
client-side PHP file which was stored in the user's web server
director, ie /var/www. I moved that file to be kept in the repository
and the user can set up his own symlinks that point to it in whatever
webserver he is using. I deleted the symlinks that were in the
repository.