Install rate limiter on creating new games
Addresses #18, which was caused by overloading the UUID generator. Each
instance of the Flask server is limited to 55 requests for a new UUID
per minute. New games that are aborted due to rate limits from the UUID
are deleted using the `del()` keyword but that should not happen in
theory. Also increased maximum number of games stored in cache at a time
to 1,000.