Make guesses and keyboard part of game state object
In this commit we change the way that the efficient keyboard map and the
guess-color map are stored.  instead of being generated on the fly on
each GET request they are kept in the game state's data dictionary. Also
changed the JSON response to no longer include unneeded fields,
including solution, progress_grid, keyboard_map, and maybe more.