Only include HTTP response code for POST methods
Previously, if you sent a valid POST method with a valid guess for a
valid game you would get the entire game status as a response. Now, you
only get the response status for the guess.

Also, if you sent a POST method for a guess that was invalid, you got a
400 error. Now, you get a 200 response code, but the error describes
what went wrong.  This will help to differentiate between problems
caused by breaking the game's rules, i.e. guess over 5 letters long, and
breaking the API's rules, ie posting a guess for a game that does not
exist.