Accept custom solutions for newly created games
The server can now change a game to have a custom solution when it
receives a POST request with a `"solution:<solution>"` key-value pair in
either Form data or JSON data.

This feature is accessible via 2 different endpoints. The first is,
`/v2/game/<game_uuid>/custom`, which is a new endpoint. It is also
accessible when a POST request sent to `/v1/game/<game_uuid>` contains a
`solution` key-value instead of a `guess` key-value. Note that a request
containing both `solution` and `guess` is rejected.